dig lookup: Fix domain-specific qtype is spread to all current domains
The dig lookup plugins allows the qtype to be set either using
the `qtype` option or by passing it as part of the domain that
is to be resolved:
    lookup('community.general.dig', 'example.com', qtype='AAAA')
    lookup('community.general.dig', 'example.com/AAAA')

However, even when using the latter one, the `qtype` variable is
set globally. That means, if there are multiple domains to be queried,
a domain with qtype included overrides the qtype for all domains
which is not expected:
    lookup('community.general.dig', 'a.example.com/a', 'b.example.com/AAAA', qtype='AAAA')

would query the AAAA records for both a.example.com and b.example.com
instead of one for the A record for a.example.com and one for the AAAA
record of b.example.com

To fix this global `qtype` variable is now used only for the `qtype`
option while the "inline" specification of qtypes that are attached
to a domain name is moved from the option-processing down to the
query call where it doesn't affect the qtype of other domains.